+++ src/main/java/org/jsoup/nodes/PseudoTextElement.java
@@ -6,7 +6,8 @@ import org.jsoup.parser.Tag;
 /**
  * Represents a {@link TextNode} as an {@link Element}, to enable text nodes to be selected with
  * the {@link org.jsoup.select.Selector} {@code :matchText} syntax.
- * @deprecated use {@link Element#selectNodes(String, Class)} instead, with selector of <code>::textnode</code> and class <code>TextNode</code>.
+ * @deprecated use {@link Element#selectNodes(String, Class)} instead, with selector of <code>::textnode</code> and class <code>TextNode</code>;
+ * will be removed in jsoup 1.24.1.
  */
 @Deprecated
 public class PseudoTextElement extends Element {