+++ src/main/java/org/jsoup/parser/HtmlTreeBuilder.java
@@ -60,7 +60,7 @@ public class HtmlTreeBuilder extends TreeBuilder {
         "button", "fieldset", "input", "keygen", "object", "output", "select", "textarea"
     };
 
-    /** @deprecated This is not used anymore. Will be removed in a future release. */
+    /** @deprecated Not used anymore; configure parser depth via {@link Parser#setMaxDepth(int)}. Will be removed in jsoup 1.24.1. */
     @Deprecated
     public static final int MaxScopeSearchDepth = 100;