+++ src/main/java/org/jsoup/select/Evaluator.java
@@ -436,7 +436,7 @@ public abstract class Evaluator {
 
         /**
          @deprecated since 1.22.1, use {@link #AttributeKeyPair(String, String)}; the previous trimQuoted parameter is no longer used.
-         This constructor will be removed in a future release.
+         This constructor will be removed in jsoup 1.24.1.
          */
         @Deprecated
         public AttributeKeyPair(String key, String value, boolean ignored) {
@@ -1046,7 +1046,7 @@ public abstract class Evaluator {
     }
 
     /**
-     @deprecated This selector is deprecated and will be removed in a future version. Migrate to <code>::textnode</code> using the <code>Element#selectNodes()</code> method instead.
+     @deprecated This selector is deprecated and will be removed in jsoup 1.24.1. Migrate to <code>::textnode</code> using the <code>Element#selectNodes()</code> method instead.
      */
     @Deprecated
     public static final class MatchText extends Evaluator {
@@ -1056,7 +1056,7 @@ public abstract class Evaluator {
             // log a deprecated error on first use; users typically won't directly construct this Evaluator and so won't otherwise get deprecation warnings
             if (!loggedError) {
                 loggedError = true;
-                System.err.println("WARNING: :matchText selector is deprecated and will be removed in a future version. Use Element#selectNodes(String, Class) with selector ::textnode and class TextNode instead.");
+                System.err.println("WARNING: :matchText selector is deprecated and will be removed in jsoup 1.24.1. Use Element#selectNodes(String, Class) with selector ::textnode and class TextNode instead.");
             }
         }